moved logic to useVenue hook. changed variable names to full ones - latitude and longitude, deleted else statement.
